updates to welcome message
